Separate error page and error shared components
