+++ src/pages/routes.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense, lazy } from "react";
+import { Suspense, lazy } from "react";
 import { Route, Switch, Redirect } from "react-router-dom";
 
 import HomePage from "pages/home-page";
@@ -7,10 +7,10 @@ import { FeaturedPage } from "pages/featured-page";
 import HallOfFamePage from "pages/hall-of-fame-page";
 import TagsPage from "pages/tags-page";
 import { SearchResultsPage } from "pages/search-results-page";
-import { ErrorBoundary, ErrorFallback } from "pages/error-handling";
-import { NoMatchPage } from "./no-match-page";
-import { Spinner } from "components/core";
-import { MonthlyRankingsPage } from "./monthly-rankings-page";
+import { ErrorFallbackPage } from "pages/error-fallback-page";
+import { NoMatchPage } from "pages/no-match-page";
+import { MonthlyRankingsPage } from "pages/monthly-rankings-page";
+import { ErrorBoundary, Spinner } from "components/core";
 
 const AsyncViewProject = lazy(() => import("pages/project-details-page"));
 const AsyncAboutPage = lazy(() => import("pages/about-page"));
@@ -18,7 +18,7 @@ const TimelinePage = lazy(() => import("pages/timeline-page"));
 
 const Routes = (props) => {
   return (
-    <ErrorBoundary fallback={<ErrorFallback />}>
+    <ErrorBoundary fallback={<ErrorFallbackPage />}>
       <Suspense fallback={<Spinner />}>
         <Switch>
           <Route exact path="/" component={HomePage} />
